+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+## [0.4.0] - 2022-04-02
+
+Add support for classic text field
+Add graphql endpoint `brsGenesisScaffolding`
+
 ## [0.3.2] - 2022-03-31
 
 Add support for toggle and rich text field